rename Main component to Landing and reroute
