+++ client/src/Components/Error404/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Error404 = () => {
+  return <h1>404. Sorry, the page you are looking for is unavailable.</h1>;
+};
+
+export default Error404;
